story media component draws the media
